Completion notification/update is delayed Fixes #7149
